+++ src/components/ProfileBadges/index.jsx
@@ -1,0 +1,10 @@
+import ProfileBadges from './ProfileBadges';
+import ProfileBadgesProvider from './ProfileBadgesProvider';
+
+const ProfileBadgesWithProvider = () => (
+  <ProfileBadgesProvider>
+    <ProfileBadges />
+  </ProfileBadgesProvider>
+);
+
+export default ProfileBadgesWithProvider;