+++ src/data/urls.js
@@ -1,3 +1,5 @@
 import { getConfig } from '@edx/frontend-platform';
 
 export const getGammaProfileUrl = (username) => `${getConfig().LMS_BASE_URL}/gamma_dashboard/api/v0/avatar-progress/${username}/`;
+
+export const getUserBadgesUrl = (username) => `${getConfig().LMS_BASE_URL}/gamma_dashboard/api/v0/user-badges/${username}/`;