+++ src/index.js
@@ -1,2 +1,3 @@
 export { default as AvatarProgress } from './components/AvatarProgress';
+export { default as ProfileBadges } from './components/ProfileBadges';
 export { HeaderUserMenuItems, LearningHeaderUserMenuItems } from './components/HeaderUserMenuItems';