+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 - [NAU-591] create widgets: HeaderUserMenuItems and LearningHeaderUserMenuItems
+- Add `ProfileBadges` widget that lists a user's earned badges (name + description) on the profile page
 
 ## Fixed
 - [OST-68] add message field to LearningHeader user menu items